# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI Agent API",
        "description": "API documentation for the AI Agent Application",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/clients": {
            "get": {
                "tags": [
                    "AI Clients"
                ],
                "summary": "Get client details",
                "description": "Retrieves details of the authenticated client",
                "operationId": "getClientDetails",
                "responses": {
                    "200": {
                        "description": "Client details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "client": {
                                            "properties": {
                                                "uuid": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "clientToken": []
                    }
                ]
            }
        },
        "/api/v1/clients/messages": {
            "get": {
                "tags": [
                    "AI Clients"
                ],
                "summary": "Get default messages",
                "description": "Retrieves default messages configured for the authenticated client",
                "operationId": "getDefaultMessages",
                "responses": {
                    "200": {
                        "description": "Default messages retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "role": {
                                                        "type": "string",
                                                        "example": "system"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "You are a helpful assistant."
                                                    },
                                                    "ordering": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "clientToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "clientToken": {
                "type": "apiKey",
                "description": "Client authentication token",
                "name": "X-AiClient-Token",
                "in": "header"
            }
        }
    },
...
}